third_party/blink/web_tests/wpt_internal/attribution-reporting/simple-aggregatable-report-with-filtering-id.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<meta name=variant content="?with-filtering-id=true&with-max-bytes=false">
<meta name=variant content="?with-filtering-id=false&with-max-bytes=false">
<meta name=variant content="?with-filtering-id=true&with-max-bytes=true">
<meta name=variant content="?with-filtering-id=false&with-max-bytes=true">
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/attribution-reporting/resources/helpers.js"></script>
<script>

// Payload with contributions [{key: 0xA85, value: 1664}], padded to 20
// contributions (and with default filtering IDs)
const DEFAULT_MAX_BYTES_DEFAULT_FILTERING_ID =
  'omRkYXRhlKNiaWRBAGV2YWx1ZUQAAAaAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAqFo2JpZEEAZX' +
  'ZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAGZidWNr' +
  'ZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAA' +
  'AAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFs' +
  'dWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldF' +
  'AAAAAAAAAAAAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAA' +
  'AACjYmlkQQBldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZU' +
  'QAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAA' +
  'AAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAK' +
  'NiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAA' +
  'AABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAGZidWNrZXRQAAAAAA' +
  'AAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAAAo2Jp' +
  'ZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAG' +
  'ZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAA' +
  'AAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAABpb3Blcm' +
  'F0aW9uaWhpc3RvZ3JhbQ==';

// Payload with contributions [{key: 0xA85, value: 1664, id: 125}], padded to 20
// contributions.
const DEFAULT_MAX_BYTES_NON_DEFAULT_FILTERING_ID =
  'omRkYXRhlKNiaWRBfWV2YWx1ZUQAAAaAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAqFo2JpZEEAZX' +
  'ZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAGZidWNr' +
  'ZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAA' +
  'AAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFs' +
  'dWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldF' +
  'AAAAAAAAAAAAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAA' +
  'AACjYmlkQQBldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZU' +
  'QAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAA' +
  'AAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAK' +
  'NiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAA' +
  'AABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAGZidWNrZXRQAAAAAA' +
  'AAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAAAAAAAAAAAAAAo2Jp' +
  'ZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAACjYmlkQQBldmFsdWVEAAAAAG' +
  'ZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRBAGV2YWx1ZUQAAAAAZmJ1Y2tldFAAAAAAAAAA' +
  'AAAAAAAAAAAAo2JpZEEAZXZhbHVlRAAAAABmYnVja2V0UAAAAAAAAAAAAAAAAAAAAABpb3Blcm' +
  'F0aW9uaWhpc3RvZ3JhbQ==';

// Payload with contributions [{key: 0xA85, value: 1664, id: 125}], padded to 20
// contributions where contributions ids are 2 bytes.
const NON_DEFAULT_MAX_BYTES_NON_DEFAULT_FILTERING_ID =
  'omRkYXRhlKNiaWRCAH1ldmFsdWVEAAAGgGZidWNrZXRQAAAAAAAAAAAAAAAAAAAKhaNiaWRCAA' +
  'BldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZi' +
  'dWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAA' +
  'AAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRC' +
  'AABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAG' +
  'ZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAA' +
  'AAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaW' +
  'RCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAA' +
  'AGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAA' +
  'AAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNi' +
  'aWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAA' +
  'AAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAA' +
  'AAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAK' +
  'NiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVE' +
  'AAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAA' +
  'AAAAAAAAAAAAAAAAAAAGlvcGVyYXRpb25paGlzdG9ncmFt';

// Payload with contributions [{key: 0xA85, value: 1664, id: 0}], padded to 20
// contributions where contributions ids are 2 bytes.
const NON_DEFAULT_MAX_BYTES_DEFAULT_FILTERING_ID =
  'omRkYXRhlKNiaWRCAABldmFsdWVEAAAGgGZidWNrZXRQAAAAAAAAAAAAAAAAAAAKhaNiaWRCAA' +
  'BldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZi' +
  'dWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAA' +
  'AAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRC' +
  'AABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAG' +
  'ZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAA' +
  'AAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaW' +
  'RCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAA' +
  'AGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAA' +
  'AAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNi' +
  'aWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAA' +
  'AAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAA' +
  'AAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAK' +
  'NiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVE' +
  'AAAAAGZidWNrZXRQAAAAAAAAAAAAAAAAAAAAAKNiaWRCAABldmFsdWVEAAAAAGZidWNrZXRQAA' +
  'AAAAAAAAAAAAAAAAAAAGlvcGVyYXRpb25paGlzdG9ncmFt';

attribution_reporting_promise_test(async t => {
  assert_equals(new Set([
    DEFAULT_MAX_BYTES_DEFAULT_FILTERING_ID,
    DEFAULT_MAX_BYTES_NON_DEFAULT_FILTERING_ID,
    NON_DEFAULT_MAX_BYTES_NON_DEFAULT_FILTERING_ID,
    NON_DEFAULT_MAX_BYTES_DEFAULT_FILTERING_ID
  ]).size, 4)

  const searchParams = new URLSearchParams(location.search);
  const withFilteringId = searchParams.get('with-filtering-id') == 'true';
  const withMaxBytes = searchParams.get('with-max-bytes') == 'true';
  let expected_agg_service_payload;
  if(withMaxBytes) {
    expected_agg_service_payload = withFilteringId
      ? NON_DEFAULT_MAX_BYTES_NON_DEFAULT_FILTERING_ID
      : NON_DEFAULT_MAX_BYTES_DEFAULT_FILTERING_ID;
  } else {
    expected_agg_service_payload = withFilteringId
      ? DEFAULT_MAX_BYTES_NON_DEFAULT_FILTERING_ID
      : DEFAULT_MAX_BYTES_DEFAULT_FILTERING_ID;
  }

  const expectedSourceDebugKey = '246';
  const expectedTriggerDebugKey = '357';

  registerAttributionSrcByImg(createRedirectChain([
    {
      cookie: attributionDebugCookie,
      source: {
        aggregation_keys: {
          geoValue: '0x5',
        },
        destination: 'https://{{host}}',
        debug_key: expectedSourceDebugKey,
      },
    },
    {
      trigger: {
        aggregatable_trigger_data: [
          {
            key_piece: '0xA80',
            source_keys: ['geoValue'],
          }
        ],
        aggregatable_values: {
          geoValue: {
            value: 1664,
            ...(withFilteringId ? { filtering_id: '125' } : {})
          },
        },
        debug_key: expectedTriggerDebugKey,
        ...(withMaxBytes ? { aggregatable_filtering_id_max_bytes : 2 } : {} ),
      },
    },
  ]));

  const payload = await pollAggregatableReports();
  assert_equals(payload.reports.length, 1);
  const report = JSON.parse(payload.reports[0].body);
  assert_equals(report.source_debug_key, expectedSourceDebugKey);
  assert_equals(report.trigger_debug_key, expectedTriggerDebugKey);
  const shared_info = JSON.parse(report.shared_info);
  assert_own_property(shared_info, 'version');
  assert_equals(shared_info.version, '1.0');
  assert_own_property(shared_info, 'debug_mode');
  assert_equals(shared_info.debug_mode, 'enabled');
  const agg_service_payload = report.aggregation_service_payloads[0];
  assert_own_property(agg_service_payload, 'debug_cleartext_payload');
  assert_equals(
    agg_service_payload.debug_cleartext_payload,
    expected_agg_service_payload
  );

  const debugPayload = await pollAttributionSuccessDebugAggregatableReports();
  assert_equals(debugPayload.reports.length, 1);
  const debugReport = JSON.parse(debugPayload.reports[0].body);
  assert_equals(debugReport.source_debug_key, expectedSourceDebugKey);
  assert_equals(debugReport.trigger_debug_key, expectedTriggerDebugKey);
  const debug_shared_info = JSON.parse(debugReport.shared_info);
  assert_own_property(debug_shared_info, 'version');
  assert_equals(debug_shared_info.version, '1.0');
  assert_own_property(debug_shared_info, 'debug_mode');
  assert_equals(debug_shared_info.debug_mode, 'enabled');
  const debug_agg_service_payload = debugReport.aggregation_service_payloads[0];
  assert_own_property(debug_agg_service_payload, 'debug_cleartext_payload');
  assert_equals(
    debug_agg_service_payload.debug_cleartext_payload,
    expected_agg_service_payload
  );

}, 'Ensure the aggregation service payload include the provided filtering id.');

</script>
